pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "babeltest"
version = "0.1.0"
description = "GraphQL for unit testing - a declarative, language-agnostic test specification language"
readme = "babeltest-spec.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Kohl", email = "kohlbern@gmail.com" }
]
keywords = ["testing", "unit-testing", "declarative", "polyglot"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "mypy>=1.0",
]

[project.scripts]
babel = "babeltest.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/babeltest"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
